# generated/openapi.json
{
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": { "type": "string" },
          "correlationId": { "format": "uuid", "type": "string" },
          "fieldErrors": {
            "additionalProperties": {
              "items": { "type": "string" },
              "type": "array"
            },
            "type": "object"
          },
          "message": { "type": "string" }
        },
        "required": ["code", "correlationId", "message"],
        "type": "object"
      },
      "HealthResponse": {
        "additionalProperties": false,
        "properties": {
          "service": { "enum": ["api"], "type": "string" },
          "status": { "enum": ["ok"], "type": "string" }
        },
        "required": ["service", "status"],
        "type": "object"
      },
      "BeginRegistration": {
        "additionalProperties": false,
        "properties": {
          "email": { "format": "email", "maxLength": 320, "type": "string" },
          "password": { "maxLength": 256, "minLength": 12, "type": "string" }
        },
        "required": ["email", "password"],
        "type": "object"
      },
      "ConfirmRegistration": {
        "additionalProperties": false,
        "properties": {
          "token": { "maxLength": 200, "minLength": 1, "type": "string" }
        },
        "required": ["token"],
        "type": "object"
      },
      "Login": {
        "additionalProperties": false,
        "properties": {
          "email": { "format": "email", "maxLength": 320, "type": "string" },
          "password": { "maxLength": 256, "minLength": 1, "type": "string" }
        },
        "required": ["email", "password"],
        "type": "object"
      },
      "Session": {
        "additionalProperties": false,
        "properties": {
          "status": { "enum": ["ENABLED", "SUSPENDED"], "type": "string" },
          "userId": { "format": "uuid", "type": "string" }
        },
        "required": ["status", "userId"],
        "type": "object"
      },
      "CreateDraftOffering": {
        "additionalProperties": false,
        "properties": {
          "categoryId": { "format": "uuid", "type": "string" },
          "slug": { "maxLength": 160, "minLength": 1, "type": "string" },
          "summary": { "maxLength": 1000, "type": "string" },
          "title": { "maxLength": 240, "minLength": 1, "type": "string" }
        },
        "required": ["categoryId", "slug", "title"],
        "type": "object"
      },
      "DraftOffering": {
        "additionalProperties": false,
        "properties": {
          "businessId": { "format": "uuid", "type": "string" },
          "categoryId": { "format": "uuid", "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "id": { "format": "uuid", "type": "string" },
          "slug": { "maxLength": 160, "minLength": 1, "type": "string" },
          "status": { "enum": ["DRAFT"], "type": "string" },
          "summary": { "maxLength": 1000, "type": ["string", "null"] },
          "title": { "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "version": { "minimum": 1, "type": "integer" }
        },
        "required": [
          "businessId",
          "categoryId",
          "createdAt",
          "id",
          "slug",
          "status",
          "summary",
          "title",
          "updatedAt",
          "version"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "V1 decision-completion marketplace HTTP API",
    "title": "Commerce Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/registrations": {
      "post": {
        "description": "Answers identically whether or not the address is already registered.",
        "operationId": "beginRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BeginRegistration" }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Registration accepted for email-control proof"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Invalid registration input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Too many registration attempts"
          }
        },
        "tags": ["Identity"]
      }
    },
    "/api/v1/auth/registrations/confirmations": {
      "post": {
        "operationId": "confirmRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmRegistration" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            },
            "description": "Account created and session established"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Registration link is invalid or has expired"
          }
        },
        "tags": ["Identity"]
      }
    },
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Login" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            },
            "description": "Session established"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Invalid credentials input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Credentials rejected"
          }
        },
        "tags": ["Identity"]
      }
    },
    "/api/v1/auth/sessions/current": {
      "delete": {
        "operationId": "logout",
        "responses": {
          "204": { "description": "Session ended" },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Request origin is missing or not allowed"
          }
        },
        "tags": ["Identity"]
      },
      "get": {
        "operationId": "getCurrentSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            },
            "description": "Current authenticated session"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "No authenticated session"
          }
        },
        "tags": ["Identity"]
      }
    },
    "/api/v1/businesses/{businessId}/offerings": {
      "post": {
        "operationId": "createDraftOffering",
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDraftOffering" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DraftOffering" }
              }
            },
            "description": "Draft Offering created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Business cannot author offerings"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Business or catalog resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Offering slug conflict"
          }
        },
        "tags": ["Offering"]
      }
    },
    "/api/v1/businesses/{businessId}/offerings/{offeringId}": {
      "get": {
        "operationId": "getDraftOffering",
        "parameters": [
          {
            "in": "path",
            "name": "businessId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "in": "path",
            "name": "offeringId",
            "required": true,
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DraftOffering" }
              }
            },
            "description": "Owned Draft Offering"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Invalid identifier"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Account is not active"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "Offering not found"
          }
        },
        "tags": ["Offering"]
      }
    },
    "/api/v1/health/live": {
      "get": {
        "operationId": "getLiveness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            },
            "description": "Service is healthy"
          }
        },
        "tags": ["Health"]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            },
            "description": "Service is healthy"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
              }
            },
            "description": "A required dependency is unavailable"
          }
        },
        "tags": ["Health"]
      }
    }
  }
}